918499166898763313
ManifestID: 918499166898763313
Date: 26 July 2018 – 17:39:47 UTC
Version: EU-279457

+++ Assembly-CSharp/NavType.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-public enum NavType : short
+public enum NavType : byte
 {
 	Floor,
 	LeftWall,
